+++ packages/web/src/components/organisms/TransactionForm/index.spec.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import { WithdrawHolderCard } from './'
+import { TransactionForm } from '.'
 import 'src/__mocks__/window/matchMedia.mock'
 
 jest.mock('src/fixtures/dev-kit/hooks')
 jest.mock('@dev/graphql')
 
-describe(`${WithdrawHolderCard.name}`, () => {
+describe(`${TransactionForm.name}`, () => {
   test('Snapshot', () => {
-    const component = render(<WithdrawHolderCard propertyAddress="propertyAddress" />)
+    const component = render(<TransactionForm propertyAddress="propertyAddress" />)
     const tree = component.baseElement
     expect(tree).toMatchSnapshot()
   })
